All Files
Uploaded all files but  the Missile does not connect to launcher , due to the Reason that i have no clue how to work with 3dsmax and what i tried with blender didn't work as intended. Maybe there is also a mistake in the Lua. I count resize the launcher but than the texture is missing . Maybe anyone can take a look and help with it